+++ changelog.md
@@ -1,3 +1,8 @@
+## H2DE-4.0.0
+#### ADDED
+- color surface
+- surface rendering
+
 ## H2DE-b-4.0.3
 #### ADDED
 - button events